services/feature-store/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
pydantic==2.5.0
merlin-core==0.8.0
merlin-models==0.12.0
merlin-systems==0.10.0
nvtabular==1.9.0
cudf-cu11==23.08.*
cupy-cuda11x==12.2.0
pandas==2.0.3
numpy==1.24.3
pyarrow==13.0.0
fastparquet==2023.8.0
redis==5.0.1
kafka-python==2.0.2
asyncpg==0.29.0
python-dotenv==1.0.0
sentry-sdk[fastapi]==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
aiofiles==23.2.1
schedule==1.2.0
protobuf==4.24.4